--- src/main/kotlin/kr/hhplus/be/server/coupon/domain/model/DiscountTarget.kt
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.coupon.domain.model
-
-import java.math.BigDecimal
-
-data class DiscountTarget(
-    val orderItemId: Long,
-    val discountAmount: BigDecimal
-)
