--- src/main/kotlin/kr/hhplus/be/server/order/application/dto/DiscountSpec.kt
@@ -1,19 +1,0 @@
-package kr.hhplus.be.server.order.application.dto
-
-import java.math.BigDecimal
-
-class DiscountSpec {
-    class Coupon {
-        data class Root(
-            val userId: Long,
-            val userCouponIds: List<Long>,
-            val items: List<Item>
-        )
-        data class Item(
-            val productId: Long,
-            val variantId: Long,
-            val quantity: Int,
-            val subTotal: BigDecimal
-        )
-    }
-}
