+++ src/main/kotlin/kr/hhplus/be/server/product/domain/Product.kt
@@ -50,10 +50,6 @@ class Product(
         variant.product = this
     }
 
-    /**
-     * 단순 수량에 대한 검증만 하는 것이 아니므로 네이밍을 purchase로 가져감.
-     * (상품이 구매 가능한 상태인지도 검증하기 때문에..)
-     */
     fun reduceStockByPurchase(variantId: Long, quantity: Int) {
         validatePurchasability(variantId, quantity)
         findVariant(variantId).reduceStock(quantity)
